+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+### Features
+
+* Download Apple App Store Sales reports
+* Parse Apple App Store Sales reports
+
 ## 1.0.0
 
 ### Features